REGRESSION (r202927): The internal size of the ImageBuffer is scaled twice by the context scaleFactor https://bugs.webkit.org/show_bug.cgi?id=159981 <rdar://problem/27429465>
Reviewed by Myles Maxfield.

Source/WebCore:

The change to propagate color spaces through ImageBuffers created an
alternate version of createCompatibleBuffer. This version accidentally
attempted to take the display resolution (i.e. hidpi) into account
when creating the buffer, which meant it was being applied twice.

The fix is simply to remove that logic. The caller of the method
will take the resolution into account, the same way they did
with the old createCompatibleBuffer method.

Test: fast/hidpi/pdf-image-scaled.html

* platform/graphics/cg/ImageBufferCG.cpp:
(WebCore::ImageBuffer::createCompatibleBuffer): Don't calculate
a resolution - just use the value of 1.0.

LayoutTests:

* fast/hidpi/pdf-image-scaled-expected.html: Added.
* fast/hidpi/pdf-image-scaled.html: Added.
* fast/hidpi/resources/circle.pdf: Added.

Identifier: 178218@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203543 268f45cc-cd09-0410-ab3c-d52691b4dbfc