AX: Video Controls: Volume cannot be adjusted using VO. https://bugs.webkit.org/show_bug.cgi?id=160107
Reviewed by Dean Jackson.

Source/WebCore:

The volume slider in video tag had 0.01 step which caused the screen reader adjusting it slowly.
Changed the step to 0.05 and added the aria-valuetext attribute to the slider, so that the value
is spoken in percentage.

Test: accessibility/mac/video-volume-slider-accessibility.html

* Modules/mediacontrols/mediaControlsApple.js:
(Controller.prototype.createControls):
(Controller.prototype.handleVolumeSliderInput):
(Controller.prototype.updateVolume):

LayoutTests:

* accessibility/mac/video-volume-slider-accessibility-expected.txt: Added.
* accessibility/mac/video-volume-slider-accessibility.html: Added.


Identifier: 178300@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203669 268f45cc-cd09-0410-ab3c-d52691b4dbfc