REGRESSION(r200931): Invalid cast in highestAncestorToWrapMarkup() https://bugs.webkit.org/show_bug.cgi?id=160163
Reviewed by Michael Catanzaro.

Since r200931 the result of enclosingNodeOfType() in highestAncestorToWrapMarkup() is downcasted to Element, but
the result of enclosingNodeOfType() can be a Node that is not an Element, in this case is Text. The cast is not
needed at all since that node is passed to editingIgnoresContent() and selectionFromContentsOfNode() and both
receive a Node not an Element.

* editing/markup.cpp:
(WebCore::highestAncestorToWrapMarkup): Remove invalid cast.

Identifier: 178313@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203683 268f45cc-cd09-0410-ab3c-d52691b4dbfc