MathOperator: Add a mapping from combining to non-combining equivalents https://bugs.webkit.org/show_bug.cgi?id=159513
Patch by Frederic Wang <fwang@igalia.com> on 2016-07-25
Reviewed by Darin Adler.

Source/WebCore:

Many math fonts provide stretch variants and assemblies for combining characters but not for
their non-combining equivalent. In the MathML recommendation, it is suggested to use
non-combining charaters, so we allow the operator stretching code to look for constructions
associated to these non-combining characters in order to still be able to stretch the
combining ones.

Test: mathml/presentation/bug159513.html

* rendering/mathml/MathOperator.cpp:
(WebCore::MathOperator::getGlyph): New function extending getBaseGlyph to retrieve the glyph
data for an arbitrary character.
(WebCore::MathOperator::getMathVariantsWithFallback): This helper function calls
getMathVariants for the base glyph. If no constructions are available, it calls
getMathVariants for the glyph associated to equivalent fallback characters as listed in the
small characterFallback table.
(WebCore::MathOperator::calculateStretchyData): Call getMathVariantsWithFallback instead of
getMathVariants. Note that we do not need to do that for calculateDisplayStyleLargeOperator
as we do not use fallback for large operators.
* rendering/mathml/MathOperator.h:
(WebCore::MathOperator::getBaseGlyph): Use getGlyph to implement this function.

LayoutTests:

* mathml/presentation/bug159513.html: Added.
* platform/gtk/mathml/presentation/bug159513-expected.png: Added.
* platform/gtk/mathml/presentation/bug159513-expected.txt: Added.
* platform/ios-simulator/TestExpectations: Skip this test on iOS.
* platform/mac/TestExpectations: Skip this test on Mac.

Identifier: 178343@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203714 268f45cc-cd09-0410-ab3c-d52691b4dbfc