[Soup] Test http/tests/xmlhttprequest/auth-reject-protection-space.html fails since added in r203743 https://bugs.webkit.org/show_bug.cgi?id=160235
Reviewed by Michael Catanzaro.

It times out in release and crashes in debug due to an ASSERT_NOT_REACHED(). The soup backend doesn't really
support rejecting the protection space and continuing with the next one, but we can at least continue with the
request without crendential to make the test pass.

* platform/network/soup/ResourceHandleSoup.cpp:
(WebCore::ResourceHandle::receivedChallengeRejection):

Identifier: 178403@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203775 268f45cc-cd09-0410-ab3c-d52691b4dbfc