[iOS] Remove unused textContentType SPI from _WKFormInputSession https://bugs.webkit.org/show_bug.cgi?id=160260
Patch by Chelsea Pugh <cpugh@apple.com> on 2016-07-27
Reviewed by Dan Bernstein.

* UIProcess/API/Cocoa/_WKFormInputSession.h: Remove unused SPI.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textInputTraits]): Set textContentType of _traits to whatever we classify it as based
on the assisted node info. The default textContentType is nil, and that is our fallback in our method
for determining textContentType.
(-[WKFormInputSession textContentType]): Deleted.
(-[WKFormInputSession setTextContentType:]): Deleted.

Identifier: 178419@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203791 268f45cc-cd09-0410-ab3c-d52691b4dbfc