Align MediaList with the CSSOM specification https://bugs.webkit.org/show_bug.cgi?id=160257
Reviewed by Ryosuke Niwa.

Source/WebCore:

Align MediaList with the CSSOM specification:
- https://www.w3.org/TR/cssom-1/#the-medialist-interface

In particular, the parameter to item() / deleteMedium() and
appendMedium() is now mandatory.

Firefox and Chrome agree with the specification.

Test: fast/css/MediaList-mediaText-null.html
      fast/css/MediaList-parameters.html

* css/MediaList.idl:

LayoutTests:

* fast/css/MediaList-mediaText-null-expected.txt: Added.
* fast/css/MediaList-mediaText-null.html: Added.
Add test coverage for MediaList.mediaText to make sure it is not nullable
and treats null as the empty string. Our IDL did not match the
specification here but our behavior was correct. Therefore, this test is
passing with and without my change. I just wanted to make sure we had
good coverage since I updated our IDL to match the specification.

* fast/css/MediaList-parameters-expected.txt: Added.
* fast/css/MediaList-parameters.html: Added.
Add test coverage for mandatory parameters.

* fast/dom/non-numeric-values-numeric-parameters-expected.txt:
* fast/dom/script-tests/non-numeric-values-numeric-parameters.js:
Update existing test to reflect behavior change.

Identifier: 178423@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203795 268f45cc-cd09-0410-ab3c-d52691b4dbfc