First parameter to SVGDocument.createEvent() should be mandatory https://bugs.webkit.org/show_bug.cgi?id=160296
Reviewed by Alex Christensen.

Source/WebCore:

First parameter to SVGDocument.createEvent() should be mandatory:
- https://www.w3.org/TR/SVG2/struct.html#InterfaceDocumentExtensions
- https://dom.spec.whatwg.org/#document

This patch drops createEvent() from SVGDocument so that the spec-compliant
Document.createEvent() is used instead. In the specification, SVGDocument
no longer exists and is a simple Document.

Firefox and Chrome agree with the specification.

Test: svg/dom/SVGDocument-createEvent.html

* svg/SVGDocument.idl:

LayoutTests:

Add test coverage.

* svg/dom/SVGDocument-createEvent-expected.txt: Added.
* svg/dom/SVGDocument-createEvent.html: Added.


Identifier: 178449@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203821 268f45cc-cd09-0410-ab3c-d52691b4dbfc