<attachment> action label doesn't wrap, but needs to for some localizations https://bugs.webkit.org/show_bug.cgi?id=160431 <rdar://problem/27526464>
Reviewed by Simon Fraser.

Test: fast/attachment/attachment-wrapping-action.html

* rendering/RenderThemeIOS.mm:
(WebCore::AttachmentInfo::buildWrappedLines):
(WebCore::AttachmentInfo::AttachmentInfo):
(WebCore::AttachmentInfo::buildTitleLines): Deleted.
Rework buildTitleLines into buildWrappedLines so that we have a
generic way to add lines that wrap up to N lines and then center-truncate.

Make use of it for the action text, and follow the same rules/width as
the title text already does.

* fast/attachment/attachment-wrapping-action-expected-mismatch.html: Added.
* fast/attachment/attachment-wrapping-action.html: Added.
Add a test that ensures that the action label wraps by filling one with
non-breaking and breaking spaces, which will wrap onto two lines, and one
with just a single space, which will always fit in one line, and ensuring
that they do not match.


Identifier: 178573@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204005 268f45cc-cd09-0410-ab3c-d52691b4dbfc