Ensure StringView lifetime is correct inside InlineTextBox https://bugs.webkit.org/show_bug.cgi?id=160584
Reviewed by Darin Adler.

A TextRun contains a StringView. Previously, we had a helper function which
could possibly create a new string, and then create a TextRun whose StringView
pointed inside this new string. Then, the string would be destroyed because it
was on the stack.

Luckily, this is benign because this new string would always share a
StringImpl with the string passed into the function (and the StringView
would point into this StringImpl). However, relying on this is dangerous
and we are trying to enforce StringView lifetime checks in
https://bugs.webkit.org/show_bug.cgi?id=160384.

The solution is to make a helper function which returns this built string
so that the caller can own this built string, thereby extending its lifetime.
Unfortunately, this built string can't be a StringView because the same
mechanism must be used for text-combine, which returns originalText(), which
may return a built string if it is a RenderCounter. Perhaps it would be
valuable to migrate RenderCounter::originalText() to something else, or to
divorce text-combine from originalText(), but that is out of scope for this
patch. Similarly, hyphenation uses this same mechanism, but because the
hyphenated string needs to be created, we must own the string storage (so
StringView is incorrect here).

Another reason is for performance: StringView::substring() is faster to use
than String::substringSharingImpl(), since it does not allocate and destroy a
StringImpl on the heap. This patch migrates away from
String::substringSharingImpl() and onto StringView::substring().

No new tests because there is no behavior change.

* rendering/InlineTextBox.cpp:
(WebCore::InlineTextBox::localSelectionRect):
(WebCore::InlineTextBox::paint):
(WebCore::InlineTextBox::selectionStartEnd):
(WebCore::InlineTextBox::paintSelection):
(WebCore::InlineTextBox::hyphenatedStringForTextRun):
(WebCore::InlineTextBox::constructTextRun):
* rendering/InlineTextBox.h:
* rendering/RenderCombineText.cpp:
(WebCore::RenderCombineText::maybeGetCombinedStringForRendering):
(WebCore::RenderCombineText::getStringToRender): Deleted.
* rendering/RenderCombineText.h:


Identifier: 178806@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204276 268f45cc-cd09-0410-ab3c-d52691b4dbfc