CharacterData.data setter optimization is not spec-compliant and is observable https://bugs.webkit.org/show_bug.cgi?id=160712
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

Rebaseline W3C test now that more checks are passing.

* web-platform-tests/dom/ranges/Range-mutations-expected.txt:

Source/WebCore:

We have an optimization in CharacterData.setData() that avoid doing any
work if the new data is identical to the old data. However, this
optimization is not spec-compliant:
- https://dom.spec.whatwg.org/#dom-characterdata-data

This is an issue because this is observable using DOM mutation observer
or ranges.

To address the issue, we now:
1. Only use the optimization if there are not DOM mutation observers /
   listeners on the page.
2. We now update the associated range when we use the fast path as well.

Test: fast/dom/Text/setData-no-op-mutation.html

* dom/CharacterData.cpp:
(WebCore::canUseSetDataOptimization):
(WebCore::CharacterData::setData):

LayoutTests:

Add test coverage to make sure that DOM Mutation observers get notified
when CharacterData.data is set to the same data.

* fast/dom/Text/setData-no-op-mutation-expected.txt: Added.
* fast/dom/Text/setData-no-op-mutation.html: Added.


Identifier: 178832@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204316 268f45cc-cd09-0410-ab3c-d52691b4dbfc